Add Parallel Scan code
Parallel implementation of a reduction operation over an array.

When given an array of n elements, a scan returns a new array that has,
in each position 1<=i<=n, the sum of all numbers between the 1st and the
i-th position of the given array.

Could be refactored in order to enable the implementation of a generic
associative operation (e.g. we then could choose multiplication in place
of sum), using the Template Method pattern.